refactor(plugins): remove oauth pass-through wrappers
